% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{check_valid_style}
\alias{check_valid_style}
\title{Is the supplied syntax a vaild syntax?}
\usage{
check_valid_style(style)
}
\arguments{
\item{style}{a style to validate that it's available.}
}
\description{
Is the supplied syntax a vaild syntax?
}

