% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{get_header}
\alias{get_header}
\title{Create Stan highlight header for LaTeX or HTML}
\usage{
get_header(formatter = c("latex", "html"))
}
\description{
Create Stan highlight header for LaTeX or HTML
}

