% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{pipe_in}
\alias{pipe_in}
\title{Build and run command with system(..., intern=TRUE)
(functions like sprintf())}
\usage{
pipe_in(cmd, input = NULL)
}
\description{
Build and run command with system(..., intern=TRUE)
(functions like sprintf())
}

