% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{set_knitr_header}
\alias{set_knitr_header}
\title{Set the opts_knit 'header' to the default knitr header combined with the
Pygments header}
\usage{
set_knitr_header()
}
\description{
\code{set_knitr_header()} will automatically get the knitr output format and
use this to create a header.
}

