% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{stanhl}
\alias{stanhl}
\title{Highlight Stan model code}
\usage{
stanhl(x, formatter = c("latex", "html"))
}
\arguments{
\item{x}{character model specification}

\item{formatter}{either "latex" or "html"; the format of output}
}
\description{
Uses Python Pygements to highly Stan model code.
}

