% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{stanhl_file}
\alias{stanhl_file}
\title{Use Python Pygements to highlight Stan model code from a file.}
\usage{
stanhl_file(file)
}
\arguments{
\item{file}{a filename to a Stan model}
}
\description{
Use Python Pygements to highlight Stan model code from a file.
}
