% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\docType{data}
\name{stanhl_opts}
\alias{stanhl_opts}
\title{Global options for stanhl}
\format{\preformatted{List of 2
 $ get:function (name)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 3 15 3 49 15 49 3 3
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f9c9a7cf908> 
 $ set:function (name, value)  
  ..- attr(*, "srcref")=Class 'srcref'  atomic [1:8] 5 15 5 61 15 61 5 5
  .. .. ..- attr(*, "srcfile")=Classes 'srcfilecopy', 'srcfile' <environment: 0x7f9c9a7cf908> 
}}
\usage{
stanhl_opts
}
\description{
Global options for stanhl
}
\keyword{datasets}

