% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{stanhl_styles}
\alias{stanhl_styles}
\alias{check_valid_style}
\alias{stanhl_opts}
\title{List all styles (including plugins) available to Pygments}
\format{
An object of class \code{list} of length 5.
}
\usage{
stanhl_styles()

check_valid_style(style)

stanhl_opts
}
\arguments{
\item{style}{a style to validate that it's available.}
}
\description{
List all styles (including plugins) available to Pygments
}
\keyword{datasets}
