% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{pipe_in}
\alias{pipe_in}
\title{Build and run command with system(..., intern=TRUE)}
\usage{
pipe_in(cmd, input = NULL)
}
\arguments{
\item{cmd}{command to execute with \code{system(....)}.}

\item{input}{string input to be passed through standard input to \code{cmd}.}
}
\description{
Build and run command with system(..., intern=TRUE)
}

