% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{stanhl}
\alias{stanhl}
\title{Highlight Stan model code}
\usage{
stanhl(x)
}
\arguments{
\item{x}{character model specification}
}
\description{
Uses Python Pygements to highly Stan model code.
}

