% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R, R/stanhl-package.R
\docType{package}
\name{stanhl}
\alias{stanhl}
\alias{stanhl-package}
\title{Highlight Stan model code}
\usage{
stanhl(x)
}
\arguments{
\item{x}{character model specification.}
}
\description{
Uses Python Pygements to highly Stan model code.

stanhl uses Python Pygments to highlight Stan code.
}

