% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanhl-package.R, R/highlight.R
\docType{package}
\name{stanhl}
\alias{stanhl}
\title{stanhl: Stan model syntax highlighting for knitr}
\usage{
stanhl(x)
}
\description{
stanhl uses Python Pygments to highlight Stan code.
}
