% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R, R/stanhl-package.R
\docType{package}
\name{stanhl}
\alias{stanhl}
\title{Use pygements to highlight Stan model code.}
\usage{
stanhl(x)
}
\arguments{
\item{x}{stan code to highlight}
}
\description{
stanhl uses Python Pygments to highlight Stan code.
}
