% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/highlight.R
\name{stanhl_file}
\alias{stanhl_file}
\title{Highlight Stan model code from file}
\usage{
stanhl_file(file)
}
\arguments{
\item{file}{a filename to a Stan model}
}
\description{
Uses Python Pygements to highly Stan model code from a file.
}

