% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\docType{data}
\name{stanhl_opts}
\alias{stanhl_opts}
\title{Global options for stanhl}
\format{
An object of class \code{list} of length 5.
}
\usage{
stanhl_opts
}
\description{
Global options for stanhl
}
\keyword{datasets}
