% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaults.R
\name{stanhl_styles}
\alias{stanhl_styles}
\alias{check_valid_style}
\alias{stanhl_opts}
\title{List all styles (including plugins) available to Pygments}
\usage{
stanhl_styles()

check_valid_style(style)

stanhl_opts(opts = list())
}
\arguments{
\item{style}{a style to validate that it's available.}

\item{opts}{Options for knitr}
}
\description{
List all styles (including plugins) available to Pygments
}
