% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/defaults.R
\name{stanhl_styles}
\alias{stanhl_styles}
\title{List all styles (including plugins) available to Pygments}
\usage{
stanhl_styles()
}
\description{
List all styles (including plugins) available to Pygments
}

